fix: close side windows should not kill background processes

+++ src-tauri/src/lib.rs
@@ -100,7 +100,7 @@ pub fn run() {
             setup_mcp(app);
             setup_sidecar(app).expect("Failed to setup sidecar");
             setup_engine_binaries(app).expect("Failed to setup engine binaries");
-            // TODO(any) need to wire up with frontend 
+            // TODO(any) need to wire up with frontend
             // let handle = app.handle().clone();
             // tauri::async_runtime::spawn(async move {
             //     handle_app_update(handle).await.unwrap();
@@ -109,11 +109,13 @@ pub fn run() {
         })
         .on_window_event(|window, event| match event {
             tauri::WindowEvent::CloseRequested { .. } => {
-                let client = Client::new();
-                let url = "http://127.0.0.1:39291/processManager/destroy";
-                let _ = client.delete(url).send();
+                if window.label() == "main" {
+                    let client = Client::new();
+                    let url = "http://127.0.0.1:39291/processManager/destroy";
+                    let _ = client.delete(url).send();
 
-                window.emit("kill-sidecar", ()).unwrap();
+                    window.emit("kill-sidecar", ()).unwrap();
+                }
             }
             _ => {}
         })
